requirements.txt
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
Pillow==9.2.0
tqdm==4.64.0
